SI-4489 feat(notes): Репозиторий для работы с Hive
